chore: move MERSUniformData
